# docs/api-audit/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tecbunny API Surface",
    "version": "1.0.0",
    "description": "Generated from repository source by scripts/audit-api-surface.mjs."
  },
  "servers": [
    {
      "url": "{{API_BASE_URL}}"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/products/scraper": {
      "options": {
        "description": "OPTIONS /api/products/scraper",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Observed or expected 200 response"
          },
          "400": {
            "description": "Observed or expected 400 response"
          },
          "401": {
            "description": "Observed or expected 401 response"
          },
          "403": {
            "description": "Observed or expected 403 response"
          },
          "404": {
            "description": "Observed or expected 404 response"
          },
          "500": {
            "description": "Observed or expected 500 response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-route-file": "apps/api/src/app/api/products/scraper/route.ts",
        "x-database-tables": [
          "products"
        ],
        "x-frontend-usage": []
      },
      "post": {
        "description": "POST /api/products/scraper",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Observed or expected 201 response"
          },
          "400": {
            "description": "Observed or expected 400 response"
          },
          "500": {
            "description": "Observed or expected 500 response"
          },
          "503": {
            "description": "Observed or expected 503 response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-route-file": "apps/api/src/app/api/products/scraper/route.ts",
        "x-database-tables": [
          "products"
        ],
        "x-frontend-usage": []
      }
    },
    "/api/products/scraper/ai": {
      "options": {
        "description": "OPTIONS /api/products/scraper/ai",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Observed or expected 200 response"
          },
          "400": {
            "description": "Observed or expected 400 response"
          },
          "401": {
            "description": "Observed or expected 401 response"
          },
          "403": {
            "description": "Observed or expected 403 response"
          },
          "404": {
            "description": "Observed or expected 404 response"
          },
          "500": {
            "description": "Observed or expected 500 response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-route-file": "apps/api/src/app/api/products/scraper/ai/route.ts",
        "x-database-tables": [],
        "x-frontend-usage": []
      },
      "post": {
        "description": "POST /api/products/scraper/ai",
        "parameters": [],
        "responses": {
          "400": {
            "description": "Observed or expected 400 response"
          },
          "500": {
            "description": "Observed or expected 500 response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-route-file": "apps/api/src/app/api/products/scraper/ai/route.ts",
        "x-database-tables": [],
        "x-frontend-usage": []
      }
    },
...
    "/api/services/engineers": {
      "get": {
        "description": "GET /api/services/engineers",
        "parameters": [],
        "responses": {
          "400": {
            "description": "Observed or expected 400 response"
          },
          "500": {
            "description": "Observed or expected 500 response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-route-file": "apps/api/src/app/api/services/engineers/route.ts",
        "x-database-tables": [],
        "x-frontend-usage": []
      },
      "post": {
        "description": "POST /api/services/engineers",
        "parameters": [],
        "responses": {
          "400": {
            "description": "Observed or expected 400 response"
          },
          "500": {
            "description": "Observed or expected 500 response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-route-file": "apps/api/src/app/api/services/engineers/route.ts",
        "x-database-tables": [],
        "x-frontend-usage": []
      }
    },
...
  }
}